plumber: Evaluate scripts for arg matches
Plumber lets us run arbitrary scripts as actions but provides a very limited set of patterns.
Patterns can check the message itself but are oblivious of the _context_ of the message, like its working directory, whether it is a source code repository, whether a certain file exists in the working directory or anywhere above it.

This change allows us to define patterns that run arbitrary scripts, evaluated just like plumb start actions, and the exit code of the scripts determines whether the rule matches the message.

A few potential use cases:

* Git and Mercurial commit identifiers look very similar. This change allows us to check what kind of repository we're in before determining how to handle the commit identifier.
* Project directories may include tags files (as generated by ctags(1), say). When plumbing something that could be an identifier, this change lets us check whether such a file is present and even whether it contains the potential identifier before committing to an action to take.
* Project directories may include dependencies, usually in a subdirectory. This change would let us check those for files in addition to the working directory, similarly to how the basic rules look for C headers in include directories.

Some of these are _possible_ without this change but involve a lot of busywork.
For git/mercurial commits, say, one would need to configure
* two rules that match commits, one for each system, each with a unique message property (destination, say) and each plumbing to a client appropriate for the corresponding system, and
* one rule that matches commits but _without_ any unique properties that plumbs to a client that determines to appropriate destination from the context (like the working directory) and then plumbs the same data _again_ but with the unique property required to trigger one of the first rules.

NOTE: This change is a proof of concept.
It does not introduce new objects or verbs but takes over the unused `arg matches` combination.
An example pattern that checks whether $wdir is a git repository looks like this:
	arg matches rc -c 'git rev-parse --git-dir >/dev/null >[2=1]'